Polymorphism
Added a different toNode method that would resize the piles depending on the type. Before every single pile was 200 pixels tall, i changed it so that the top (final, get, draw) piles are 150 pixels tall and the normal piles are 500 px tall. This means there is a lot more room to play on the board